// src/integration/api-types.ts
export interface components {
    schemas: {
// ...
        GemeentePakket: {
            pakketId?: string;
            /** @example Zaaksysteem */
            pakketNaam?: string;
            /** @example zaaksysteem */
            pakketSlug?: string;
            /** @example 3.2 */
            versie?: string;
            /** @example productie */
            status?: string;
            leverancier?: {
                /** @example Centric */
                naam?: string;
            };
            referentiecomponenten?: {
                naam?: string;
                guid?: string;
            }[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}